Fixed broken tests following mocked Clone and more
This stuff is a bit tricky. The take away is that RunData.clone()
copies tests and assignments folders from course.CloneDir() (it is
set using the QUICKFEED_REPOSITORY_PATH) and it is problematic to
change this to something else for some of the tests. For example,
the TestRunTests is designed to fetch some example tests from the
qf101 course and run those tests to inspect that the scoring
works.
