+++ ci/docker_test.go
@@ -121,7 +121,7 @@ func TestDockerBindDir(t *testing.T) {
 
 	const (
 		script  = `ls /quickfeed`
-		wantOut = "Dockerfile\nrun.sh\ntests\n" // content of testdata (or /quickfeed inside the container)
+		wantOut = "Dockerfile\nassignments\nrun.sh\ntests\n" // content of testdata (or /quickfeed inside the container)
 		image   = "golang:latest"
 	)
 	docker, closeFn := dockerClient(t)
