+++ ci/run_tests.go
@@ -3,13 +3,12 @@ package ci
 import (
 	"context"
 	"fmt"
-	"io"
-	"io/fs"
 	"os"
 	"path/filepath"
 	"strings"
 	"time"
 
+	"github.com/quickfeed/quickfeed/internal/fileop"
 	"github.com/quickfeed/quickfeed/internal/rand"
 	"github.com/quickfeed/quickfeed/kit/score"
 	"github.com/quickfeed/quickfeed/qf"
@@ -132,64 +131,10 @@ func (r RunData) clone(ctx context.Context, sc scm.SCM, dstDir string) error {
 			return err
 		}
 	}
-
 	// Copy the tests and assignment repos to the destination directory
-	err = copyDir(testsDir, dstDir)
+	err = fileop.CopyDir(testsDir, dstDir)
 	if err != nil {
 		return err
 	}
-	return copyDir(assignmentDir, dstDir)
-}
-
-// Copy directory from src directory to dst, ignoring .git directories.
-func copyDir(src, dst string) error {
-	dst = filepath.Join(dst, filepath.Base(src))
-	return filepath.WalkDir(src, func(path string, d fs.DirEntry, err error) error {
-		if err != nil {
-			return err
-		}
-		if d.IsDir() {
-			return nil
-		}
-		if strings.Contains(path, ".git") {
-			return nil
-		}
-		rel, err := filepath.Rel(src, path)
-		if err != nil {
-			return err
-		}
-		dstPath := filepath.Join(dst, rel)
-		if err := os.MkdirAll(filepath.Dir(dstPath), 0o700); err != nil {
-			return err
-		}
-		return copyFile(path, dstPath)
-	})
-}
-
-// Copy file from src to dst.
-func copyFile(srcFile, dstFile string) (err error) {
-	in, err := os.Open(srcFile)
-	if err != nil {
-		return err
-	}
-	defer func() {
-		closeErr := in.Close()
-		if err == nil {
-			err = closeErr
-		}
-	}()
-
-	out, err := os.Create(dstFile)
-	if err != nil {
-		return err
-	}
-	defer func() {
-		closeErr := out.Close()
-		if err == nil {
-			err = closeErr
-		}
-	}()
-
-	_, err = io.Copy(out, in)
-	return
+	return fileop.CopyDir(assignmentDir, dstDir)
 }
