+++ scm/mock_test.go
@@ -14,6 +14,7 @@ import (
 )
 
 func TestMockClone(t *testing.T) {
+	dstDir := t.TempDir()
 	s := scm.NewMockSCMClient()
 	ctx := context.Background()
 	cloneTests := []struct {
@@ -27,9 +28,9 @@ func TestMockClone(t *testing.T) {
 			opt: &scm.CloneOptions{
 				Organization: qtest.MockOrg,
 				Repository:   qf.StudentRepoName("user"),
-				DestDir:      "testdata",
+				DestDir:      dstDir,
 			},
-			wantPath: filepath.Join("testdata", "user-labs"),
+			wantPath: filepath.Join(dstDir, "user-labs"),
 			wantErr:  false,
 		},
 		{
@@ -37,9 +38,9 @@ func TestMockClone(t *testing.T) {
 			opt: &scm.CloneOptions{
 				Organization: qtest.MockOrg,
 				Repository:   qf.AssignmentRepo,
-				DestDir:      "testdata",
+				DestDir:      dstDir,
 			},
-			wantPath: filepath.Join("testdata", "assignments"),
+			wantPath: filepath.Join(dstDir, "assignments"),
 			wantErr:  false,
 		},
 		{
@@ -47,16 +48,16 @@ func TestMockClone(t *testing.T) {
 			opt: &scm.CloneOptions{
 				Organization: qtest.MockOrg,
 				Repository:   qf.TestsRepo,
-				DestDir:      "testdata",
+				DestDir:      dstDir,
 			},
-			wantPath: filepath.Join("testdata", "tests"),
+			wantPath: filepath.Join(dstDir, "tests"),
 			wantErr:  false,
 		},
 		{
 			name: "missing organization info",
 			opt: &scm.CloneOptions{
 				Repository: qf.StudentRepoName("user"),
-				DestDir:    "testdata",
+				DestDir:    dstDir,
 			},
 			wantPath: "",
 			wantErr:  true,
