+++ scm/testdata/user-labs/lab1/dummy.go
@@ -1,0 +1,1 @@
+package lab1
